Move gateway stuff into its own folder
