+++ src/helpers/gateway/receive.rs
@@ -1,0 +1,81 @@
+use crate::{
+    helpers::{
+        buffers::UnorderedReceiver, gateway::wrapper::Wrapper, ChannelId, Error, Message,
+        Transport, TransportImpl,
+    },
+    protocol::RecordId,
+};
+use dashmap::DashMap;
+use futures::Stream;
+use generic_array::GenericArray;
+use std::marker::PhantomData;
+
+/// Receiving end end of the gateway channel.
+pub struct ReceivingEnd<M: Message> {
+    unordered_rx: UR<TransportImpl>,
+    _phantom: PhantomData<M>,
+}
+
+/// Receiving channels, indexed by (role, step).
+pub(super) struct GatewayReceivers<T: Transport> {
+    inner: DashMap<ChannelId, UR<T>>,
+}
+
+pub(super) type UR<T> = UnorderedReceiver<
+    <T as Transport>::RecordsStream,
+    <<T as Transport>::RecordsStream as Stream>::Item,
+>;
+
+impl<M: Message> ReceivingEnd<M> {
+    pub(super) fn new(rx: UR<TransportImpl>) -> Self {
+        Self {
+            unordered_rx: rx,
+            _phantom: PhantomData,
+        }
+    }
+
+    /// Receive message associated with the given record id. This method does not return until
+    /// message is actually received and deserialized.
+    ///
+    /// ## Errors
+    /// Returns an error if receiving fails
+    ///
+    /// ## Panics
+    /// This will panic if message size does not fit into 8 bytes and it somehow got serialized
+    /// and sent to this helper.
+    pub async fn receive(&self, record_id: RecordId) -> Result<M, Error> {
+        // TODO: proper error handling
+        let v = self.unordered_rx.recv::<Wrapper, _>(record_id).await?;
+
+        let mut buf = GenericArray::default();
+        let sz = buf.len();
+
+        buf.copy_from_slice(&v.0[..sz]);
+        Ok(M::deserialize(&buf))
+    }
+}
+
+impl<T: Transport> Default for GatewayReceivers<T> {
+    fn default() -> Self {
+        Self {
+            inner: DashMap::default(),
+        }
+    }
+}
+
+impl<T: Transport> GatewayReceivers<T> {
+    pub fn get_or_create<M: Message, F: FnOnce() -> UR<T>>(
+        &self,
+        channel_id: &ChannelId,
+        ctr: F,
+    ) -> UR<T> {
+        let receivers = &self.inner;
+        if let Some(recv) = receivers.get(channel_id) {
+            recv.clone()
+        } else {
+            let stream = ctr();
+            receivers.insert(channel_id.clone(), stream.clone());
+            stream
+        }
+    }
+}
